// packages/dukascopy-node/src/types.ts
export type Timeframe = 'tick' | 'm1' | 'm30' | 'h1' | 'd1' | 'mn1';

export type PriceType = 'bid' | 'ask';

export interface HistoryConfig {
  instrument: Instrument;
  dates: {
    from: string;
    to: string;
  };
  timeframe?: Timeframe;
  priceType?: PriceType;
  utcOffset?: number;
  volumes?: boolean;
  ignoreFlats?: boolean;
}